python-code/requirements.txt
# 核心依赖
python-binance==1.0.19
pandas==1.5.3
numpy==1.24.3
matplotlib==3.7.1
plotly==5.14.1
websocket-client==1.5.1

# 异步处理
asyncio-mqtt==0.13.0
aiohttp==3.8.4

# 数据分析和技术指标
ta-lib==0.4.25
scipy==1.10.1

# 配置和日志
python-dotenv==1.0.0
colorlog==6.7.0

# Web界面 (可选)
streamlit==1.23.1
dash==2.10.1

# 数据库 (可选)
sqlite3

# 测试
pytest==7.3.1
pytest-asyncio==0.21.0
websockets>=11.0.0
flask-socketio>=5.3.0